pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cascade-bot-mcp"
version = "0.2.0"
description = "Multi-agent Plan→Implement→Review loop with MCP + Telegram interfaces."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Christof Treitges" }]
dependencies = [
  "anthropic>=0.40",
  "ollama>=0.4",
  "openai>=1.50",
  "python-telegram-bot[ext]>=21.6",
  "aiosqlite>=0.20",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "mcp>=1.2",
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
]

[project.scripts]
cascade = "cascade.core:cli_main"
cascade-bot = "bot:main"
cascade-mcp = "mcp_server:main"

[tool.setuptools]
packages = [
  "cascade",
  "cascade.agents",
  "cascade.bot",
  "cascade.bot.handlers",
]
py-modules = ["bot", "mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
